+++ src/org/mmbase/datatypes/processors/ChainedProcessor.java
@@ -16,13 +16,13 @@ import java.util.*;
  * Chains a bunch of other processors into one new processor.
  *
  * @author Michiel Meeuwissen
- * @version $Id: ChainedProcessor.java,v 1.2 2005-11-04 23:11:52 michiel Exp $
+ * @version $Id: ChainedProcessor.java,v 1.3 2005-12-10 14:33:36 michiel Exp $
  * @since MMBase-1.7
  */
 
 public class ChainedProcessor implements Processor {
 
-    private static final int serialVersionUID = 1;
+    private static final long serialVersionUID = 1L;
 
     private List processors = new ArrayList();
