+++ src/org/mmbase/datatypes/processors/LastModified.java
@@ -18,14 +18,14 @@ import java.util.Date;
  * thrown (in other words, the field is read only).
  *
  * @author Michiel Meeuwissen
- * @version $Id: LastModified.java,v 1.2 2005-11-04 23:11:52 michiel Exp $
+ * @version $Id: LastModified.java,v 1.3 2005-12-10 14:33:36 michiel Exp $
  * @since MMBase-1.8
  * @see   LastModifier
  */
 
 public class LastModified implements CommitProcessor {
 
-    private static final int serialVersionUID = 1;
+    private static final long serialVersionUID = 1L;
 
     public void commit(Node node, Field field) {
         node.setValueWithoutProcess(field.getName(), new Date());
