+++ src/org/mmbase/datatypes/processors/Readonly.java
@@ -16,12 +16,12 @@ import org.mmbase.bridge.*;
  * If the field is set an exception is thrown (in other words, the field is read only).
  *
  * @author Nico Klasens
- * @version $Id: Readonly.java,v 1.1 2005-11-04 15:55:26 nklasens Exp $
+ * @version $Id: Readonly.java,v 1.2 2005-12-10 14:33:36 michiel Exp $
  * @since MMBase-1.8
  */
 public class Readonly implements Processor {
 
-    private static final int serialVersionUID = 1;
+    private static final long serialVersionUID = 1L;
 
     /**
      * You can plug this in on every set-action besides 'object' which will make this
@@ -30,5 +30,5 @@ public class Readonly implements Processor {
     public Object process(Node node, Field field, Object value) {
         throw new BridgeException("You cannot change the field " + field.getName());
     }
-    
+
 }
