+++ src/org/mmbase/datatypes/processors/xml/FieldSetString.java
@@ -16,19 +16,19 @@ import org.mmbase.util.*;
 /**
  * This class implements the `get' for XML field with specialization 'field'. They look like ASCII,
  * if accessed by setString, getString.
- * 
+ *
  * This can be used if the XML features of a certain field are only used in certain extensions only,
  * but must look like String in the base class. So, it allows you to have the db-type XML without
  * noticing it too much.
  *
  * @author Michiel Meeuwissen
- * @version $Id: FieldSetString.java,v 1.1 2005-10-25 12:30:26 michiel Exp $
+ * @version $Id: FieldSetString.java,v 1.2 2005-12-10 14:33:36 michiel Exp $
  * @since MMBase-1.8
  */
 
 public class FieldSetString implements  Processor {
     private static final Logger log = Logging.getLoggerInstance(FieldSetString.class);
-    private static final int serialVersionUID = 1;
+    private static final long serialVersionUID = 1L;
 
     protected static final String PREF = "<field><![CDATA[";
     protected static final String POST = "]]></field>";
@@ -37,5 +37,5 @@ public class FieldSetString implements  Processor {
         log.debug("Getting " + field + " from " + node + " as a String");
         return Casting.toXML(PREF + Casting.toString(value) + POST);
     }
-  
+
 }
