+++ src/org/mmbase/datatypes/processors/xml/HtmlGetString.java
@@ -14,13 +14,13 @@ import org.mmbase.bridge.*;
 /**
  * Currently is like FieldGetString.
  * @author Michiel Meeuwissen
- * @version $Id: HtmlGetString.java,v 1.1 2005-10-25 12:30:26 michiel Exp $
+ * @version $Id: HtmlGetString.java,v 1.2 2005-12-10 14:33:36 michiel Exp $
  * @since MMBase-1.8
  */
 
 public class HtmlGetString implements  Processor {
 
-    private static final int serialVersionUID = 1;
+    private static final long serialVersionUID = 1L;
 
     private Processor processor = new FieldGetString();
